Task 1
This commit allows the user to enter tasks. When on the main page simply ask the chatbot to create task and it will open the new window.

This window allows the user to create a new task. It includes task name, description and optionally a date. By Default each task status is Not Started.

This commit also includes the Code to manipulate tasks. This includes displaying tasks, deleting tasks and setting tasks to completed.

This can be done by any variation of show tasks or display tasks.