Task 2: Cybersecurity Mini-Game (Quiz)
This commit contains code for a mini quiz hosted by the chatbot.

This Quiz contains 15 questions, 5 of which are True/False, the last 10 are Multiple Choice.

The chatbot tells the user if they are correct or incorrect. The program then explains the answer of the question.

This Quiz tracks the correct questions and displays the result of the quiz at the end.