Task 3: Natural Language Processing (NLP) Simulation (GUI)
This commit contains the code for the NLP section of this POE.

This adds more functionality to the chatbot by increasing its list of keywords in regards to Task creations.

Instead of simply opening the task window, a task can be automatically added to the window before showing the user for final evaluation.