Fix Error with Missing Text File
The text file containing the questions for the quiz was not added to GitHub on the last commit. This moves the Text file to the "Resource" folder inside the project. The project is then told to copy the file from the resource folder if it is newer than the currently used file for the quiz. This should ensure that the text file is present in every version of the project cloned from GitHub in the future.