Final Commit
This commit contains code and updates.

This commit includes code to search for a task name and allows the user to edit details about the given task.

This commit includes code to set a reminder on the device that the reminder or task is created on.

This commit lastly includes a new button on the task window to show all saved tasks, the user can also ask to "Show tasks" in the main chat window to access this window.