Part 3 Requirements Update.
This commit contains the changes that where introduced in the updated POE Part 3 Requirements document.

The "Show more" feature for the activity log feature now works as intended. It will show the last 5 entries as requested. The user can then ask to "Show more" and the next 5 (or less) will be shown.

Creating, Editing, Deleting and Marking a task as "Complete" will now result in a log entry.

The Quiz now has a custom completion message depending if the user has a correct score below six (6) between six (6) and twelve (12) and above twelve (12).