Read ME YouTube Update
This commit adds the new YouTube video to the read me file. This new video includes the audio from myself as well as a face cam. This video goes over how to use all the new features in the POE part 3. This does not explain the code line by line as requested in the new POE Requirements given to us 22 hours after the submission opened.